requirements-dev.txt
-r requirements.txt

pytest
pytest-django
ruff
black
pre-commit
